Explicit boxing and unboxing
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1504729 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/math3/exception/MultiDimensionMismatchException.java
@@ -79,13 +79,13 @@ public class MultiDimensionMismatchException extends MathIllegalArgumentExceptio
      * @return the wrong dimension stored at {@code index}.
      */
     public int getWrongDimension(int index) {
-        return wrong[index];
+        return wrong[index].intValue();
     }
     /**
      * @param index Dimension index.
      * @return the expected dimension stored at {@code index}.
      */
     public int getExpectedDimension(int index) {
-        return expected[index];
+        return expected[index].intValue();
     }
 }
